+++ RELEASE_NOTES.md
@@ -6,10 +6,11 @@
 - Added hint for the backglass server settings that fozzy matching is not supported by the VPin Studio. You can enable it, but the Studio will always only look for exact .directb2s matches.
 - Added "Statistics" tab to the Table Data Manager dialog.
 - Added Table Filters.
+- Added Backglass Manager.
 
 <img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/cards/transparent-cards.png" width="600" />
 
 ### Bugfixes
 
-- Table Data Manager: Fixed missing information about unsupported ROM names.
+- Table Data Manager: Fixed miscellaneous issues with the info badges.
 - Pause Menu: Fixed video author allow list so that an empty list shows all videos available in the pause menu.